# docs/swagger.yaml
definitions:
...
  job_rate_status.CreateRequest:
    properties:
      class:
        type: string
      label:
        type: string
      order:
        type: integer
    type: object
  job_rate_status.JobRateStatus:
    properties:
      class:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      label:
        type: string
      order:
        type: integer
      updatedAt:
        type: string
    type: object
  job_rate_status.UpdateRequest:
    properties:
      class:
        type: string
      label:
        type: string
      order:
        type: integer
    type: object
  job_resident.CreateRequest:
    properties:
      email:
        type: string
      homePhone:
        type: string
      mobilePhone:
        type: string
      name:
        type: string
    type: object
  job_resident.JobResident:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      homePhone:
        type: string
      id:
        type: integer
      jobId:
        type: integer
      mobilePhone:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  job_resident.ListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/job_resident.JobResident'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  job_resident.UpdateRequest:
    properties:
      email:
        type: string
      homePhone:
        type: string
      mobilePhone:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8090
info:
  contact:
    email: support@jvair.com
    name: API Support
    url: http://www.jvair.com/support
  description: API para el sistema JVAIR
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: JVAIR API
  version: "1.0"
paths:
...
  /api/v1/job-rate-statuses:
    get:
      consumes:
      - application/json
      description: Get all job rate statuses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/job_rate_status.JobRateStatus'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List job rate statuses
      tags:
      - JobRateStatuses
    post:
      consumes:
      - application/json
      description: Create a new job rate status
      parameters:
      - description: Job rate status data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/job_rate_status.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/job_rate_status.JobRateStatus'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create job rate status
      tags:
      - JobRateStatuses
  /api/v1/job-rate-statuses/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a job rate status by ID
      parameters:
      - description: Job Rate Status ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete job rate status
      tags:
      - JobRateStatuses
    get:
      consumes:
      - application/json
      description: Get a job rate status by ID
      parameters:
      - description: Job Rate Status ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/job_rate_status.JobRateStatus'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get job rate status by ID
      tags:
      - JobRateStatuses
    put:
      consumes:
      - application/json
      description: Update a job rate status by ID
      parameters:
      - description: Job Rate Status ID
        in: path
        name: id
        required: true
        type: integer
      - description: Job rate status data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/job_rate_status.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/job_rate_status.JobRateStatus'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update job rate status
      tags:
      - JobRateStatuses
...
  /api/v1/jobs/{jobId}/residents:
    get:
      consumes:
      - application/json
      description: Get all residents for a specific job
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: integer
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/job_resident.ListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List job residents
      tags:
      - JobResidents
    post:
      consumes:
      - application/json
      description: Create a new resident contact for a job
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: integer
      - description: Resident data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/job_resident.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/job_resident.JobResident'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create job resident
      tags:
      - JobResidents
  /api/v1/jobs/{jobId}/residents/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a resident by ID
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: integer
      - description: Resident ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete job resident
      tags:
      - JobResidents
    put:
      consumes:
      - application/json
      description: Update a resident by ID
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: integer
      - description: Resident ID
        in: path
        name: id
        required: true
        type: integer
      - description: Resident data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/job_resident.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/job_resident.JobResident'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update job resident
      tags:
      - JobResidents
...
securityDefinitions:
  BearerAuth:
    description: Ingrese el token con el prefijo 'Bearer '
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...